Improve encryption support
